refactor: 필드명, 데이터타입 수정, 상태전이 가드 추가 (#11)

+++ src/main/java/com/realdev/readle/domain/quiz/entity/QuizQuestion.java
@@ -43,7 +43,7 @@ public class QuizQuestion {
   private QuestionType questionType;
 
   @Column(name = "order_no", nullable = false)
-  private short orderNo;
+  private Integer orderNo;
 
   @Lob
   @Column(name = "question_text", nullable = false)
